Added Debugging ability
There is now a property in config.js that contains a boolean on whether to display console debugging information.

There is a function in core that can be called in other js files called consoleDebug that checks the value of this property and if true then sends the info to the console. This way your static files can be the same between dev/test/production witht the only changes being in config.js