Add TailWind CSS
The Vista demo uses TailWind CSS to style their page. As we are about to copy the header, adding the same css is the quickest way to implement it.
There is ability to optimize this later by removing unneeded classes, or making custom classes for the header, or switching entirely to tailwind.